src/PostApi/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
httpx>=0.24.1
grpcio>=1.54.0
grpcio-tools>=1.54.0
PyJWT>=2.6.0
googleapis-common-protos>=1.56.4
protobuf>=4.21.6,<5.0.0
